Replace createRef with useRef
When a key is held down, the app tends to crash, saying that the audio
element is not defined. I think it's because a new ref is created every
time the drum pad rerenders, and somehow the keydown event handler fires
before the ref to the audio element is set. Idk.

Though I don't think it ever crashed before I introduced local state to
the drum pad.

I guess the takeaway is to use `useRef` inside functional components.